+++ gossipsub-interop/run.py
@@ -8,8 +8,7 @@ import subprocess
 from network_graph import generate_graph
 import experiment
 
-from analyze_message_deliveries import analyse_message_deliveries
-from analyze_message_dessemination_rate import analyse_message_dessemination_rate
+from analyze_message_deliveries import analyse_message_deliveries, analyse_message_dessemination_rate
 
 params_file_name = "params.json"
